requirements.txt
streamlit==1.30.0
  openai==1.97.1
  langchain==0.2.7
  langchain-openai==0.1.10
  langchain-community==0.2.7
  langchain-core==0.2.13
  langchain-text-splitters==0.2.2
  langsmith==0.1.77
  langgraph==0.0.67
  numpy==1.26.4
  pandas==2.2.2
  scikit-learn==1.5.0
  nltk==3.8.1
  textblob==0.17.1
  requests==2.31.0
  beautifulsoup4==4.12.3
  matplotlib==3.8.4
  seaborn==0.13.2
  tensorflow==2.16.1
  torch==2.3.0
  pillow==10.2.0
  python-dotenv==1.0.1
  language-tool-python
  textstat
  pyspellchecker
  vaderSentiment
  streamlit-chat